refactor: 중복 BaseEntity 제거
