--- src/main/java/hanium/modic/backend/common/baseEntity/BaseEntity.java
@@ -1,27 +1,0 @@
-package hanium.modic.backend.common.baseEntity;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@MappedSuperclass
-@NoArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-
-	@CreatedDate
-	@Column(updatable = false)
-	private LocalDateTime createAt;
-
-	@LastModifiedDate
-	private LocalDateTime updateAt;
-}
